Move away from SteerableActors; introduce MovableActors
At the risk of replicating functionality seen in the Gdx.Ai package, we
are implementing a new way to move Actors, through the use of different
WeightedGoals.

It will be the responsibility of the Actor/Agent to act when one of the
goals is reached, or if something happens along the way. All that the
GoalMovement class does is collaborate with the actor to determine its
“speed” vector.

The system was built to be “hot-swappable” with other movement
techniques, such as the old SteerableImage.

The movement is ultimately less natural than with steering behaviors
but, for the time being, it is easier to code, predict and understand