Refactor pheromone code so that it is mostly bound together by a Pheromone class
Lots of implicit conventions are now in place, in particular regarding
certain static methods that should be called. E.g., the
initializePheromones() method in SimplisticAnt can and should be called
before the world is created, so that upon world creation, all currently
existing pheromones are prompted to add themselves to the world layers.

Overall, this change was quite drastic, but needed to maintain some
sense of organization in the code