Establish a general framework for Pheromone Controllers
Now Pheromones can be controlled with a PheromoneController. This
interface has an update() method which returns whether or not the
control action was taken. A typical example is the simple
IncreasePheromoneController which always increases the pheromone by a
given amount (thus always returning true).

Two other controllers used to establish the logical connections are the
ConditionalIncreasePheromoneController, which acts as the
PheromoneController based on a functional condition, and the
AlternativeOrderedPheromoneController, which uses the Composite design
pattern to stop once the first of its Controllers is met.