Introduce the NumberProducer concept
A NumberProducer is a functional interface that can produce numbers.
Typical examples are random number generations given by normal or
uniform distributions. On the other hand, a simple NumberProducer is
simply constant, implemented via “() -> NUMBER” (e.g. for a
NumberProducer that always returns 3, do () -> r).

This is meant to be used in deeper parts of the engine as a set of
configurable parameters of the population, and also to add more
randomness. For instance, what if we want the “RampageForFood” state to
happen differently for different ants? We can use a NumberProducer for
that.

It would be great if java allowed us to implicitly convert any number
to a ConstantNumberProducer, but oh well