Adjust collision detection (and A*) to allow for "Flexible Bounding Boxes"
This technique uses information abound the object coverage in a given
cell to adjust the BoundingBox coverage. For instance, even if an agent
is in an object’s bounding box, it may be considered as “not colliding”
if the coverage of that bounding box in that cel is very small. This
value is currently hardcoded at 30% but should be changed in the future