Implement an initial spreadDIrtAtCell
While ants dig, they will spread a given percentage of dirt to
neighboring cells. This will come from a source (i.e. spreadSource)
cell. The ideal way to do this would be to actually keep track of which
cells have agents and deny throwing dirt at that. I’ll probably code
that in the future.